Move shader registration to a separate class
